+++ client/src/components/Button.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
+import styles from './styles/Button.css';
 
 const Button = props => (
-  <button type="button" className="qualified-button">Get pre-qualified</button>
+  <button type="button" className={styles.qualifiedButton}>Get pre-qualified</button>
 );
 
 export default Button;
