+++ client/src/components/Chart.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
+import styles from './styles/Chart.css';
 
 const Chart = props => (
-  <svg width="100%" height="100%" viewBox="0 0 42 42" className="donut">
+  <svg width="80%" height="80%" viewBox="0 0 42 42" className={styles.donut}>
     <circle className="donut-hole" cx="21" cy="21" r="15.91549430918954" fill="#fff"/>
     <circle className="donut-ring" cx="21" cy="21" r="15.91549430918954" fill="transparent"
             stroke="#d9f0f9" strokeWidth="3"/>
@@ -21,11 +22,11 @@ const Chart = props => (
             stroke="#1b5098" strokeWidth="3"
             strokeDasharray={props.items.PMI + " " + (100 - props.items.PMI)}
             strokeDashoffset={(100 - (props.items.PI + props.items.insurance + props.items.taxes)) + 25}/>
-    <g className="chart-text">
-      <text x="50%" y="50%" className="chart-label">
+    <g className={styles.chartText}>
+      <text x="50%" y="50%" className={styles.chartLabel}>
         YOUR PAYMENT
       </text>
-      <text x="50%" y="50%" className="chart-number">
+      <text x="50%" y="50%" className={styles.chartNumber}>
         {props.monthlyPayment}
       </text>
     </g>
