+++ client/src/components/MortgageSection.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
+import styles from './styles/MortgageSection.css';
 
 const MortgageSection = props => (
-  <div className="title" onClick={props.onClick}>
-    <span className="collapsible"><b>Mortgage</b></span>
-    <i className="fas fa-angle-up" id="icon-angle"/>
+  <div className={styles.title} onClick={props.onClick}>
+    <span className={styles.collapsible}><b>Mortgage</b></span>
+    <i className="fas fa-angle-up" id={styles.iconAngle}/>
   </div>
 );
