+++ client/src/components/SubSection.jsx
@@ -3,18 +3,19 @@ import PaymentList from './PaymentList.jsx';
 import Chart from './Chart.jsx';
 import Form from'./Form.jsx';
 import Button from'./Button.jsx';
+import styles from './styles/SubSection.css';
 
 const SubSection = props => (
-  <div className="grid-container">
-    <div className="grid-item">
+  <div className={styles.gridContainer}>
+    <div className={styles.gridItem}>
       <Form item={props.items} onChangeHandler={props.onChangeHandler} onClick={props.onClick} selectedValue={props.selectedValue} onSelectChange={props.onSelectChange} checked={props.checked} onCheckHandler={props.onCheckHandler} />
     </div>
-    <div className="grid-item">
-      <div className="sub-grid-container">
-        <div className="grid-item-sm">
+    <div className={styles.gridItem}>
+      <div className={styles.subGridContainer}>
+        <div className={styles.gridItemSm}>
           <PaymentList items={props.payments} checked={props.checked}/>
         </div>
-        <div className="grid-item-sm">
+        <div className={styles.gridItemSm}>
           <Chart items={props.paymentsPercentage} monthlyPayment={props.payments.monthlyPayment} checked={props.checked}/>
         </div>
       </div>
@@ -24,3 +25,4 @@ const SubSection = props => (
 );
 
 export default SubSection;
+
