+++ client/src/components/styles/App.css
@@ -1,0 +1,5 @@
+.inner{
+    width: 50%;
+    margin: 0 auto;
+    position: relative;
+}
