+++ client/src/components/styles/Chart.css
@@ -1,0 +1,28 @@
+.chartText {
+    font: 12px "Open Sans", sans-serif;
+    fill: #000;
+    -moz-transform: translateY(0.10em);
+    -ms-transform: translateY(0.10em);
+    -webkit-transform: translateY(0.10em);
+    transform: translateY(0.10em);
+}
+
+.chartLabel{
+    font-size: 0.2em;
+    line-height: 1.4;
+    text-anchor: middle;
+    -moz-transform: translateY(-0.25em);
+    -ms-transform: translateY(-0.25em);
+    -webkit-transform: translateY(-0.25em);
+    transform: translateY(-0.25em);
+}
+
+.chartNumber {
+    font-size: 0.4em;
+    text-transform: uppercase;
+    text-anchor: middle;
+    -moz-transform: translateY(0.7em);
+    -ms-transform: translateY(0.7em);
+    -webkit-transform: translateY(0.7em);
+    transform: translateY(0.7em);
+}
