+++ client/src/components/styles/Form.css
@@ -1,0 +1,217 @@
+.form{
+    color: #444;
+    width: 100%;
+    padding: 0;
+}
+
+.input{
+    display: inline-block;
+    height: 20px;
+    padding: 2px;
+    color: #444;
+    width: 80%;
+}
+
+.inputMd{
+    display: inline-block;
+    height: 20px;
+    padding: 2px;
+    color: #444;
+    width: 80%;
+}
+
+
+.selectInput{
+    display: block;
+    color: #444;
+    background: transparent;
+    height: 25px;
+    margin: 4px 0 15px 0;
+    padding: 0;
+    width: 100%;
+}
+
+.form input[type = 'checkbox']{
+    display: inline-block;
+    width: 13px;
+    height: 13px;
+    padding: 0;
+    margin: 0 4px 0 0;
+    vertical-align: bottom;
+    position: relative;
+    top: -1px;
+    *overflow: hidden;
+    cursor: pointer;
+    outline: 1px solid #444;
+    -moz-appearance:none;
+    -webkit-appearance:none;
+}
+
+.form input[type = 'checkbox']:hover{
+    outline: 1px solid #0074e4;
+}
+
+.form input[type='checkbox']:checked:after{
+    content: "\2713";
+    color: #444;
+    text-align: center;
+}
+
+.form .check{
+    display: inline-block;
+    padding-left: 15px;
+    text-indent: -15px;
+}
+
+.form .check:hover{
+    color: #0074e4;
+}
+
+.checkboxDiv{
+    margin: 4px 0 15px 0;
+}
+
+.amountInput{
+    display: block;
+    box-shadow:0 0 0 1px rgba(68, 68, 68, 0.5);
+    padding: 0;
+    margin: 4px 0 15px 0;
+    color: #444;
+}
+.amountInput input {
+    border: 0;
+}
+
+.lgSmInput{
+    margin: 4px 0 15px 0;
+}
+
+.doubleAmountInputLg{
+    display: inline-block;
+    box-shadow:0 0 0 1px rgba(68, 68, 68, 0.5);
+    padding: 0;
+    color: #444;
+    width: 60%;
+}
+
+.doubleAmountInputSm{
+    display: inline-block;
+    box-shadow:0 0 0 1px rgba(68, 68, 68, 0.5);
+    padding: 0;
+    color: #444;
+    width: 40%;
+}
+
+.inputSm{
+    width: 80%;
+    border: 0;
+    padding: 2px;
+    height: 20px;
+    color: #444;
+}
+
+.inputLg{
+    width: 80%;
+    border: 0;
+    padding: 2px;
+    height: 20px;
+    color: #444;
+}
+
+.form label{
+    display: block;
+}
+
+.doubleInput{
+    display: block;
+    padding-left: 15px;
+    margin-left: -15px;
+}
+
+.form a{
+    margin-left: 10px;
+    display: inline-block;
+}
+
+.checkInput{
+    cursor: pointer;
+}
+
+.doubleInput a{
+    margin-left: 20px;
+}
+
+.help {
+    position: relative;
+    display: inline-block;
+    cursor: pointer;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+}
+
+.help .popuptext {
+    visibility: hidden;
+    width: 250px;
+    background-color: #444;
+    color: #fff;
+    line-height: 1.3em;
+    text-align: left;
+    border-radius: 6px;
+    padding: 8px;
+    position: absolute;
+    z-index: 1;
+    /*top: -73px;*/
+    top: 50%;
+    transform: translateY(-50%);
+    left: 105%;
+    margin-left: 10px;
+}
+
+.help .popuptext::after {
+    content: " ";
+    position: absolute;
+    top: 50%;
+    right: 100%;
+    margin-top: -5px;
+    border-width: 5px;
+    border-style: solid;
+    border-color: transparent #444 transparent transparent;
+}
+
+.help .show {
+    visibility: visible;
+    -webkit-animation: fadeIn 1s;
+    animation: fadeIn 1s;
+}
+
+@-webkit-keyframes fadeIn {
+    from {opacity: 0;}
+    to {opacity: 1;}
+}
+
+@keyframes fadeIn {
+    from {opacity: 0;}
+    to {opacity:1 ;}
+}
+
+/*.amountInput span{*/
+    /*float: right;*/
+    /*padding: 4px;*/
+/*}*/
+
+.symbols{
+    float: right;
+    padding: 4px 2px 2px 4px;
+}
+
+.rateLink{
+    float: right;
+    text-decoration: none;
+    color: #0074e4;
+}
+
+.rateLink:hover{
+    color: #62aef7;
+}
